# api/cmd/gotalk/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/room/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create room",
                "tags": [
                    "Room"
                ],
                "summary": "CreateRoom",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Room"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/room/{roomId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get detail of room",
                "tags": [
                    "Room"
                ],
                "summary": "GetRoomDetail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room Id",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/RoomDetail"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/room/{roomId}/chat": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get chat of room",
                "tags": [
                    "Room",
                    "Chat"
                ],
                "summary": "GetChat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room Id",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Chat"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send message to a chat.",
                "tags": [
                    "Room",
                    "Chat"
                ],
                "summary": "SendChatMessage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Room Id",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Send chat message request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SendChatMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/SendChatMessageResponse"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create user.",
                "tags": [
                    "User"
                ],
                "summary": "CreateUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get  current authenticated user.",
                "tags": [
                    "User"
                ],
                "summary": "GetCurrentUser",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Chat": {
            "type": "object",
            "required": [
                "history",
                "id",
                "roomId"
            ],
            "properties": {
                "history": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ChatMessage"
                    }
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                }
            }
        },
        "ChatMessage": {
            "type": "object",
            "required": [
                "isCurrentUser",
                "sentAt",
                "text",
                "user"
            ],
            "properties": {
                "isCurrentUser": {
                    "type": "boolean"
                },
                "sentAt": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/User"
                }
            }
        },
        "Error": {
            "type": "object",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "Room": {
            "type": "object",
            "required": [
                "id",
                "isPrivate",
                "ownerId"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "isPrivate": {
                    "type": "boolean"
                },
                "ownerId": {
                    "type": "string"
                }
            }
        },
        "RoomDetail": {
            "type": "object",
            "required": [
                "id",
                "isPrivate",
                "owner",
                "participants"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "isPrivate": {
                    "type": "boolean"
                },
                "owner": {
                    "$ref": "#/definitions/User"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/User"
                    }
                }
            }
        },
        "SendChatMessage": {
            "type": "object",
            "required": [
                "text"
            ],
            "properties": {
                "text": {
                    "type": "string",
                    "maxLength": 1500,
                    "minLength": 1,
                    "example": "hello world"
                }
            }
        },
        "SendChatMessageResponse": {
            "type": "object",
            "required": [
                "isOk"
            ],
            "properties": {
                "isOk": {
                    "type": "boolean"
                }
            }
        },
        "User": {
            "type": "object",
            "required": [
                "email",
                "id",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}